Split file, create functions for each case
